+++ core/src/main/java/org/elasticsearch/index/mapper/internal/RoutingFieldMapper.java
@@ -33,9 +33,8 @@ import org.elasticsearch.index.mapper.Mapper;
 import org.elasticsearch.index.mapper.MapperParsingException;
 import org.elasticsearch.index.mapper.MergeMappingException;
 import org.elasticsearch.index.mapper.MergeResult;
-import org.elasticsearch.index.mapper.ParseContext;
 import org.elasticsearch.index.mapper.MetadataFieldMapper;
-import org.elasticsearch.index.mapper.core.AbstractFieldMapper;
+import org.elasticsearch.index.mapper.ParseContext;
 
 import java.io.IOException;
 import java.util.Iterator;
@@ -53,7 +52,7 @@ public class RoutingFieldMapper extends MetadataFieldMapper {
     public static final String NAME = "_routing";
     public static final String CONTENT_TYPE = "_routing";
 
-    public static class Defaults extends AbstractFieldMapper.Defaults {
+    public static class Defaults {
         public static final String NAME = "_routing";
 
         public static final MappedFieldType FIELD_TYPE = new RoutingFieldType();
